make examples consistent
it makes more sense to first create a custom graph and then pass to the difference
class vs needing to make a custom difference class
