+++ examples/dwarf/array-example.py
@@ -2,14 +2,14 @@ __author__ = "Vanessa Sochat"
 __copyright__ = "Copyright 2022, Vanessa Sochat"
 __license__ = "MPL 2.0"
 
-from model import DwarfDifference
-
+from compspec.runner import Difference
 import os
 import sys
 import json
 
 here = os.path.abspath(os.path.dirname(__file__))
 sys.path.insert(0, here)
+from model import DwarfGraph
 
 
 def main():
@@ -21,8 +21,11 @@ def main():
         if not os.path.exists(lib):
             sys.exit(f"{lib} does not exist.")
 
-    # Run the diff!
-    runner = DwarfDifference(lib1, lib2)
+    # Create the two graphs
+    A = DwarfGraph(lib1)
+    B = DwarfGraph(lib2)
+
+    runner = Difference(A, B, "A", "B")
     result = runner.run()
     print(json.dumps(result, indent=4))
