+++ examples/python/tensorflow-example.py
@@ -40,5 +40,6 @@ def main():
             # Don't parse further nested, just a demo
             break
 
+
 if __name__ == "__main__":
     main()
